backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
gunicorn==20.1.0
SQLAlchemy==2.0.35
pyodbc==5.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
python-multipart==0.0.12